# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Action": {
        "properties": {
          "action": {
            "minLength": 1,
            "pattern": "APPROVE|REJECT|WITHDRAW|COMMENT|TRANSFER|ADD_SIGN",
            "type": "string"
          },
          "comment": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "targetUserId": {
            "format": "int64",
            "type": "integer"
          },
          "taskId": {
            "format": "int64",
            "type": "integer"
          },
          "values": {
            "additionalProperties": {},
            "type": "object"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "action",
          "version"
        ],
        "type": "object"
      },
...
      "ApiResponseListTemplate": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Template"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponseObject": {
        "properties": {
          "data": {},
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "data": {},
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "message",
          "success"
        ],
        "type": "object"
      },
...
      "Simulation": {
        "properties": {
          "applicantId": {
            "format": "int64",
            "type": "integer"
          },
          "schema": {
            "$ref": "#/components/schemas/Spec"
          },
          "values": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "schema"
        ],
        "type": "object"
      },
      "Spec": {
        "properties": {
          "allowRepeatApproval": {
            "type": "boolean"
          },
          "allowSelfApproval": {
            "type": "boolean"
          },
          "allowWithdraw": {
            "type": "boolean"
          },
          "applicantIds": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array",
            "uniqueItems": true
          },
          "applicantType": {
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/Field"
            },
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/Node"
            },
            "type": "array"
          },
          "startNodeId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Template": {
        "properties": {
          "description": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "schema": {
            "$ref": "#/components/schemas/Spec"
          }
        },
        "type": "object"
      },
...
      "WorkflowDraft": {
        "properties": {
          "businessType": {
            "pattern": "GENERAL|CONTENT",
            "type": "string"
          },
          "categoryId": {
            "format": "int64",
            "type": "integer"
          },
          "code": {
            "minLength": 1,
            "pattern": "[A-Za-z0-9_-]{2,64}",
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "schema": {
            "$ref": "#/components/schemas/Spec"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "businessType",
          "categoryId",
          "code",
          "enabled",
          "name"
        ],
        "type": "object"
      },
      "WorkflowSubmission": {
        "properties": {
          "businessId": {
            "format": "int64",
            "type": "integer"
          },
          "businessRevisionId": {
            "format": "int64",
            "type": "integer"
          },
          "businessVersion": {
            "format": "int64",
            "type": "integer"
          },
          "definitionId": {
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "maxLength": 160,
            "minLength": 0,
            "type": "string"
          },
          "values": {
            "additionalProperties": {},
            "type": "object"
          },
          "versionId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "definitionId",
          "title",
          "versionId"
        ],
        "type": "object"
      },
      "WorkflowVersionRequest": {
        "properties": {
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "通过滑块验证与登录接口取得的一次性会话令牌",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "统一响应、动作权限、数据范围与乐观锁均由服务端执行。关闭模块不开放接口。",
    "title": "Mayday API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/operations/requests": {
      "get": {
        "operationId": "WorkflowController_requests",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "box",
            "required": false,
            "schema": {
              "default": "mine",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      },
      "post": {
        "operationId": "WorkflowController_submit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/requests/{id}": {
      "get": {
        "operationId": "WorkflowController_request",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/requests/{id}/decision": {
      "post": {
        "operationId": "WorkflowController_decide",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Action"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/requests/{id}/events": {
      "get": {
        "operationId": "WorkflowController_events",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
...
    "/api/operations/requests/{id}/history": {
      "get": {
        "operationId": "WorkflowController_history",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/requests/{id}/remind": {
      "post": {
        "operationId": "WorkflowController_remind",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/requests/{id}/retry-notifications": {
      "post": {
        "operationId": "WorkflowController_retry",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
...
    "/api/operations/workflows": {
      "get": {
        "operationId": "WorkflowController_definitions",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "categoryId",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 10,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      },
      "post": {
        "operationId": "WorkflowController_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/workflows/options": {
      "get": {
        "operationId": "WorkflowController_options",
        "parameters": [
          {
            "in": "query",
            "name": "businessType",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/workflows/roles": {
      "get": {
        "operationId": "WorkflowController_roleOptions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/workflows/simulate": {
      "post": {
        "operationId": "WorkflowController_simulate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Simulation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/workflows/templates": {
      "get": {
        "operationId": "WorkflowController_templates",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListTemplate"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/workflows/{id}": {
      "delete": {
        "operationId": "WorkflowController_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      },
      "get": {
        "operationId": "WorkflowController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      },
      "put": {
        "operationId": "WorkflowController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/workflows/{id}/publish": {
      "post": {
        "operationId": "WorkflowController_publish",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
    "/api/operations/workflows/{id}/versions": {
      "get": {
        "operationId": "WorkflowController_versions",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "workflow-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}